day-14: complete challenge 2
must confess i looked up some guidance on this solution. ironically,
the solution to the problem i was facing (ignoring building the string as
it is an unnecessary performance overhead), was the appraoch i
mistakenly took in the previous day's challenge (where i assumed the
'intermediary' data structure was not necessary for the overall
solution and could be omitted).

lesson learnt: focus on what you actually need to output.